Historias de BD funcionales 100% con validaciones, optimización, entre otros

+++ database/migrations/2022_01_18_144240_create_maderas_table.php
@@ -16,8 +16,8 @@ class CreateMaderasTable extends Migration
         Schema::create('maderas', function (Blueprint $table) {
             $table->unsignedBigInteger('producto_id')->primary();
             $table->foreign('producto_id')->references('id')->on('productos')->onDelete('cascade')->onUpdate('cascade');
-            $table->float('alto');
-            $table->float('ancho');
+            $table->integer('alto');
+            $table->integer('ancho');
             $table->float('largo');
             $table->string('tipo_madera');
             $table->string('tratamiento');
